+++ ExcalidrawZ/ExcalidrawView/Engine/ExcalidrawCore/ExcalidrawCore.swift
@@ -30,6 +30,7 @@ class ExcalidrawCore: NSObject, ObservableObject {
     internal var publishError: (_ error: Error) -> Void
     var webView: ExcalidrawWebView = .init(frame: .zero, configuration: .init()) { _ in }
     lazy var webActor = ExcalidrawWebActor(coordinator: self)
+    lazy var nativeEyeDropper = ExcalidrawNativeEyeDropperCoordinator(core: self)
     
     override init() {
         self.publishError = { error in }